bot-detection-demo/app/pyproject.toml
[project]
name = "bot-detection-demo"
version = "0.1.0"
description = "Bot Detection Signup Demo - Databricks App"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "databricks-sdk>=0.30.0",
]
